Lv.4 API로깅 및 Security Config add

+++ src/main/java/org/example/expert/domain/comment/controller/CommentAdminController.java
@@ -1,10 +1,16 @@
 package org.example.expert.domain.comment.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.example.expert.domain.comment.service.CommentAdminService;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.time.LocalDateTime;
 
 @RestController
 @RequiredArgsConstructor
@@ -13,7 +19,17 @@ public class CommentAdminController {
     private final CommentAdminService commentAdminService;
 
     @DeleteMapping("/admin/comments/{commentId}")
-    public void deleteComment(@PathVariable long commentId) {
+    public void deleteComment(@PathVariable long commentId,
+                              HttpServletRequest request
+    ) {
+        Object roleAttr = request.getAttribute("userRole");
+        String role = roleAttr != null ? roleAttr.toString() : null;
+        if(!"ADMIN".equals(role)){
+            throw   new RuntimeException("권한 없음");
+        }
+
+        System.out.println("[로그]"+ LocalDateTime.now()+"commentId: "+commentId);
+
         commentAdminService.deleteComment(commentId);
     }
 }
